Exposed search to the session object.
This will allow the caller to get the row/column of a match at the
offset.  This also exposes the signals from the HistorySearch object so
that offests can be performed.

Internally the String is converted to a QRegExp so QRegExp style regular
expressions can be used for searching the history.